Modify animation loop to only depend on time
Modify FantifoClient's tick() method to not count frames since this was
a problem when the tab lost focus. Inactive tabs don't call their
requestAnimationFrame callback queue which meant that tick() had to
catch up when the tab recieved focus again. Now it calculates the
current frame and skips to that frame when the tab receives focus which
hopefully solves the problem.